1 e 2

+++ src/main/java/br/edu/ifpb/poo/Model/Aluno.java
@@ -11,16 +11,15 @@ import lombok.NoArgsConstructor;
 public class Aluno {
 
     private String nome;
-    private int matricula;
+    private long matricula;
     private List<Inscricao> inscricoes;
 
-    public Aluno(String nome, int matricula) {
+    public Aluno(String nome, long matricula) { 
         this.nome = nome;
         this.matricula = matricula;
         this.inscricoes = new ArrayList<>();
     }
 
-    //to-do: Verificar duplicidade
     public void adicionarInscricao(Inscricao insc) {
         inscricoes.add(insc);
     }
